task uploaded

+++ routes/web.php
@@ -1,36 +1,35 @@
 <?php
 
 use App\Http\Controllers\AdminController;
+use App\Http\Controllers\Auth\LoginController;
 use App\Http\Controllers\ClientController;
 use App\Http\Controllers\CompanyController;
 use App\Http\Controllers\ProductController;
+use App\Http\Middleware\Check;
 use Illuminate\Support\Facades\Route;
 
-Route::get('/', function () {
-    return view('welcome');
-});
+Route::get('/', [AdminController::class, 'index'])->name('admin.index')->middleware(['auth', Check::class . ':admin,editor,creator,deleter']);
 
-Route::get('/', [AdminController::class, 'admin']);
+Route::get('login', [LoginController::class, 'showLoginForm'])->name('login');
+Route::post('login', [LoginController::class, 'login']);
+Route::get('register', [LoginController::class, 'showRegisterForm'])->name('register');
+Route::post('register', [LoginController::class, 'register']);
+Route::get('logout', [LoginController::class, 'logout'])->name('logout');
 
+Route::get('/company', [CompanyController::class, 'index'])->name('company.index')->middleware(['auth', Check::class . ':admin,editor,creator,deleter']);
+Route::get('/company-create', [CompanyController::class, 'create'])->name('company.create')->middleware(['auth', Check::class . ':admin,creator']);
+Route::post('/company', [CompanyController::class, 'store'])->middleware(['auth', Check::class . ':admin,creator']);
+Route::delete('/company/{id}', [CompanyController::class, 'delete'])->middleware(['auth', Check::class . ':admin,deleter']); 
+Route::get('/company/{id}', [CompanyController::class, 'addProduct'])->middleware(['auth', Check::class . ':admin,editor,creator']);
 
-Route::get('/company', [CompanyController::class, 'company']);
-Route::get('/company-create', [CompanyController::class, 'create']);
-Route::post('/company', [CompanyController::class, 'store']);
-Route::delete('/company/{id}', [CompanyController::class, 'delete']);
-Route::get('/company/{id}', [CompanyController::class, 'addProduct']);
+Route::get('/product', [ProductController::class, 'product'])->name('product.index')->middleware(['auth', Check::class . ':admin,editor,creator,deleter']);
+Route::get('/product-create', [ProductController::class, 'create'])->middleware(['auth', Check::class . ':admin,creator']); 
+Route::post('/product', [ProductController::class, 'store'])->middleware(['auth', Check::class . ':admin,creator']);
+Route::delete('/product/{id}', [ProductController::class, 'delete'])->middleware(['auth', Check::class . ':admin,deleter']); 
+Route::put('/product/{id}', [ProductController::class, 'update'])->name('product.update')->middleware(['auth', Check::class . ':admin,editor']);
 
-
-Route::get('/product', [ProductController::class, 'product'])->name('product.index');
-Route::get('/product-create', [ProductController::class, 'create']);
-Route::post('/product', [ProductController::class, 'store']);
-Route::delete('/product/{id}', [ProductController::class, 'delete']);
-Route::put('/product/{id}', [ProductController::class, 'update'])->name('product.update');
-
-
-
-Route::get('/client', [ClientController::class, 'client'])->name('client.index');
-Route::get('/client-create', [ClientController::class, 'create']);
-Route::post('/client', [ClientController::class, 'store']);
-Route::delete('/client/{id}', [ClientController::class, 'delete'])->name('client.destroy');
-Route::put('/client/{user}', [ClientController::class, 'update'])->name('client.update');
-
+Route::get('/client', [ClientController::class, 'client'])->name('client.index')->middleware(['auth', Check::class . ':admin,editor,creator,deleter']);
+Route::get('/client-create', [ClientController::class, 'create'])->middleware(['auth', Check::class . ':admin,creator']);
+Route::post('/client', [ClientController::class, 'store'])->middleware(['auth', Check::class . ':admin,creator']);
+Route::delete('/client/{id}', [ClientController::class, 'delete'])->name('client.destroy')->middleware(['auth', Check::class . ':admin,deleter']);
+Route::put('/client/{user}', [ClientController::class, 'update'])->name('client.update')->middleware(['auth', Check::class . ':admin,editor']);
